refactor: clean/org ryo3-serde
