+++ crates/ryo3-serde/src/lib.rs
@@ -4,13 +4,9 @@ mod any_repr;
 mod constants;
 mod errors;
 mod macro_rules;
-mod py_serialize;
-mod rytypes;
 
-#[expect(clippy::inline_always)]
-mod safe_impl;
 pub mod ser;
 mod type_cache;
 
 pub(crate) use constants::{Depth, MAX_DEPTH};
-pub use py_serialize::SerializePyAny;
+pub use ser::py_serialize::SerializePyAny;
