+++ crates/ryo3-serde/src/ser/mod.rs
@@ -1,7 +1,12 @@
 //! Serialize python objects
 //!
 //! Currently pre-reorg
-
 mod context;
 
+pub(crate) mod py_serialize;
+mod rytypes;
+#[expect(clippy::inline_always)]
+pub(crate) mod safe_impl;
+
 pub(crate) use context::PySerializeContext;
+pub use py_serialize::SerializePyAny;
