+++ crates/ryo3-serde/src/ser/py_serialize.rs
@@ -7,15 +7,15 @@ use serde::ser::{Error as SerError, Serialize, Serializer};
 
 use crate::any_repr::any_repr;
 use crate::errors::pyerr2sererr;
+use crate::ser::PySerializeContext;
 #[cfg(feature = "ry")]
-use crate::rytypes;
-use crate::safe_impl::{
+use crate::ser::rytypes;
+use crate::ser::safe_impl::{
     SerializePyBool, SerializePyBytesLike, SerializePyDataclass, SerializePyDate,
     SerializePyDateTime, SerializePyDict, SerializePyFloat, SerializePyFrozenSet, SerializePyInt,
     SerializePyList, SerializePyMapping, SerializePyNone, SerializePySequence, SerializePySet,
     SerializePyStr, SerializePyTime, SerializePyTimeDelta, SerializePyTuple, SerializePyUuid,
 };
-use crate::ser::PySerializeContext;
 use crate::type_cache::{PyObType, PyTypeCache};
 use crate::{Depth, MAX_DEPTH};
 use pyo3::types::{PyAnyMethods, PyDict, PyMapping, PySequence};
