+++ crates/ryo3-serde/src/ser/safe_impl/mod.rs
