+++ crates/ryo3-serde/src/ser/safe_impl/py_byteslike.rs
